assets: add a DCR backend
This adds a DCR backend for gathering information about UTXOs and
signatures. Also DEXAsset, a common interface for asset backends, which
the DCR backend implements. Includes both offline and live unit tests.

Adds a DEXTx interface for getting transaction information and using
it to verify swap contracts and spend outputs.

Adds P2SH support and multi-sig. Also adds parsing of stake versions
of supported script types so that vote and revocation outputs can be
used as inputs after maturation.

* p2sh swap contract auditing

These changes modify the DCR backend constructor to be the same as
the BTC constructor. The choice of network is based on a DEX supplied
ID, and other configuration settings are loaded from a specified
filepath.

Adds support for Edwards and Schnorr signatures.