[order,matcher]: Define the Order type and begin matcher
Deviations from spec, updates to which are in other commits:
 - Hash function is blake256, not sha256.
 - Market orders do not have time-in-force, only limit orders do.

market/order:
 - Create the Order interface under the server/market/order package.
 - Create the OrderID type.
 - Implement order serialization and order ID calculation for cancel,
   market, and limit orders.
 - Define a UTXO interface for providing the UTXO data field of the
   orders.  This may change.
 - Price rate is atoms of quote asset, applied per 1e8 units of the base
   asset. Add matcher.BaseToQuote and matcher.QuoteToBase to
   perform the big integer math with an integer rate.
- NOTE: Checking the quote asset quantity of market buy orders against
   lot size,best sell order rate, and the fractional market buy buffer should
   occur only when validating incoming orders, not in the matching process.
 - 88.1% test coverage of the order package.

matcher:
 - Add sortQueue and shuffleQueue functions, to be used in order matching.
 - Define a barebones Matcher type.
 - Implement the (*Matcher).Match method and the unexported helper functions.
 - Define the Booker interface for interacting with the order book.
 - 90.7% test coverage of the matcher package

account: Primarily a placeholder, but create an AccountID type with a constructor to double-hash a pubkey as per spec.

account/pki: Primarily a placeholder.  Just define PrivateKey in terms of secp256k1, and pub/priv key size consts similarly.

market: Entirely a placeholder. Create a stub Marke type with comments outlining the primary intended functions. Hint at the need for a concrete EpochQueue type, which should either implement some interface defined by the matcher package, or have a primitive field of type []order.Order that can be provided as a concrete type to the matcher.

Use blue oak license headers in code.