asset: add a backend for Bitcoin
Adds an asset backend for bitcoin. The implementation is basically a
modified version of the DCR backend, with some additional functionality
that should allow the backend to be reused for at least some subset of
bitcoin forks. The client is btcd/rpcclient, but testing has been
against bitcoin core. Live tests are implemented via exported functions
that can be used for testing clone compatibility as well.

I've added a litecoin backend to demonstrate the clone usage. The
litecoin backend, without tests and comments, is < 40 loc.

* move extractScriptAddr script selection to caller

* move all auth to PaysToPubkey. fix blockCache reorg

* return error from SpendsUTXO

* drop BTC prefix from exports.