book: export and copy functions
Add Orders, OrdersN, ExtractN, Copy, and Realloc methods to OrderPQ

Add (*Book).SellOrders, SellOrdersN, BuyOrders, and BuyOrdersN

Make OrderPQ capacity a constructor arg, and add Realloc method.

order: cache the OrderID (slice and string formats)

book: remove OrderPricer from OrderPQ, using LimitOrder directly

Instead of an OrderPricer interface, use *order.LimitOrder inside of an
orderEntry.  This means a stub order cannot be used, so modify the
OrderPQ tests to make randome LimitOrders.

In switching OrderPQ to use a LimitOrder instead of an interface, it
exposed the need to cache order IDs (both the byte array and the string
representation).  These cached values are added to Prefix, and the
ID() and UID() methods for each of CancelOrder, MarketOrder, and
LimitOrder are modified to first check the cache before computing
the ID from the type's serialization.

//nolint:structcheck is used in the Prefix because it falsely identifies
the id and uid fields as unused, even though they are used in the
order types that embed Prefix.