server: add communication hub
Adds a websocket and http server as part of the new comms package.
The server is currently only used for websockets at the /ws endpoint,
but adding more endpoints is trivial.

Websockets are done with gorilla. I'm using the go-chi package for
http routing, which seems like overkill for a single endpoint, but
will become more useful as the number grows.

The server is pretty basic, but it does have small middleware stack
and the ability to temporarily blacklist an IP address.