spec: change json-rpc to custom message protocol
* drop json-rpc for custom json protocol

JSON-RPC is far from ideal for bi-directional websocket communcation,
and has features which make it undesireable in general for our needs,
e.g. variable-type IDs. The version field included with every request
is not needed once a websocket connection is established. There is
also some unneeded overhead for notifications that can potentially
be eliminated if we break from a strict adherence to JSON-RPC.

The proposed messaging protocol here attempts to capture the strengths
of JSON-RPC, but with a structure more suited to the needs of the DEX.