add order commitments
Desribes process for creating preimages and commitments. Adds the
fields to the appropriate JSON structures and serializations.
Describe new DEX-originating preimage request that occurs at the
end of the epoch, before sorting. Sets a DEX constant timeout for
preimage responses of 5 seconds. Inserts new rule at rule 1, making
it a conduct violation to not respond to a preimage request.

Updates the shuffling algorithm to use preimages.

Adds epoch index and epoch duration fields to the order structures.