market: epoch queue processing
This implements the basic order processing pipeline in the Market. This includes:

- cycling the epoch
- receiving, stamping, and including new orders in the appropriate epoch
- notifying book update subscribers of all epoch and book changes.
- storing new matches with the archivist

DEXArchivist now includes MatchArchiver, a new interface implemented by server/db/driver/pg for storing, updating, and retrieving matches. See this self-review for open issues on how to handle unexpected or fatal DB failures.

There is much more work on Market to work with the dex, but this is the start of epoch processing and match storage.

Market controls: Stop, Start, WaitForShutdown

Match DB.

msgjson: include both epoch index and duration in both OrderResult and trade Prefix

EpochID is index and duration.  In pg epochID is a string "idx:dur".
Include EpochID in Match, MatchSet, MatchData

order: panic on ID if Time is not set (zero value Time)