fee fixes
This commit and other changes in this PR change the way fees are
evaluated on the server. It was meant to be a couple of small changes
as I reworked the DEXAsset interface, but it became a whole thing.
In particular, the BTC and clone backends use vsize for fee evaluation
instead of byte size, and the DCR backends evaluate the fees as
zero for coinbase transactions. The second change does little for us
in production, since we don't evaluate fees on funding transactions,
but for live testing, it helps to tame some unwanted behavior.