multi: switch to coin id
This completes the switch to coin ID proposed in #75. Coin IDs are
used to identify spendable value, and are represented as a
variable-length byte array (`[]byte` in Go). Asset backends and wallets
have some latitude to choose encoding, though for UTXO-based
assets, the encoding `<32-byte tx hash>|<4-byte vout>` is recommended.

This commit and other changes in this PR change the way fees are
evaluated on the server. It was meant to be a couple of small changes
as I reworked the DEXAsset interface, but it became a whole thing.
In particular, the BTC and clone backends use vsize for fee evaluation
instead of byte size, and the DCR backends evaluate the fees as
zero for coinbase transactions. The second change does little for us
in production, since we don't evaluate fees on funding transactions,
but for live testing, it helps to tame some unwanted behavior.