client: add a BTC exchange wallet
* add a BTC wallet

A wallet interface is developed along with the wallet functionality
needed for BTC, and likely a few clones. To avoid the need to run
a btcd node or to rewrite an RPC client from scratch, a compromise
is made to use btcd's rcpclient.Client for all RPC calls, but with
wallet-related calls routed through the RawRequest method and
parsed independently.

The wallet interface offered here is intended to be somewhat
abstract, with language such as "utxos", "txid", and "vout" traded
for more generalized terms such as "coins", "id", and "index". I
believe this makes the interface more flexible and keeps the client
implementation independent of the underlying blockchain specifics.