client app skeleton
This is a basic skeleton for a client app. It enables running an HTTP
server or an RPC server. They both use the same client core, from the
new core module. The default behavior is to run the servers from a
simple but expandable TUI, but they can also be run without the TUI via
command-line arguments or configuration file. Placeholder packages for the
actual servers are in place, but don't do anything yet. A basic TUI
infrastructure is built on tview, with enough code to demonstrate its use,
but ultimately no functionality yet. I'm hoping this provides a good
framework on which concurrent development can proceed.