book: change pq book map to key on OrderID
Add (*OrderPQ).HaveOrder.
Add (*Book).HaveOrder

market: (*Market).Cancelable

Add epochOrders field to Market.  This is somewhat inefficient since
the currentEpoch in runEpochs has the same map. However, with the
current design, there is also nextEpoch to consider, so epochOrders is
updated in processOrder and processEpoch. In the future, the Epoch type
can provide a HaveOrder method, while Market will have to keep the
epochs as fields with a mutex guarding them, but it's not worth it now.