client: add a web server
A new webserver package enables client access via browser. Node is used
for compiling and bundling, but there are no production JS
dependencies, which are a major source of security concern.

It's a single-page app that borrows ideas from turbolinks and
stimulus to load and bind to new content without a new page request,
which would require a new websocket connection.

An initial market page with a chart is implemented to demonstrate
use.