spec: upgrades headers. add subsection links to readme
Adds select subsection links beneath their section link on the landing 
page. Also repositions the table of contents for each section below the 
section header. Since we paginated, we are able to upgrade the section 
headers 1 level. This enables one move level of table of contents 
navigation, though it's not used yet.

Moves the message-rate, timestamp, and coin ID subsections to a new 
encoding subsection on the comms page.