market, book: rework duplicat order detection
Duplicate order IDs in processOrder cannot really exist unless the
server time stamp is equal along with all the other serializable data.
A proper duplicate order check may be done via the commitment;
this will be in another set of changes.

In the present changes, we now enforce that multiple cancel orders
targeting the same order cannot be added to the epoch queue. If not,
ErrDuplicateCancelOrder.
Add EpochQueue.CancelTargets to facilitate this check.

Further, verify that the target order is on the books or in the epoch
queue, and that the account of the CancelOrder is the same as the
account of the target order.  If not, ErrInvalidCancelOrder.
Add (*Market).CancelableBy to perform the account check.

book: Add (*OrderPQ).Order and (*Book).Order to facilitate the above
Market changes.