client authorization
Adds handling of 'connect' route, which is now tied to the user
signing in to the client with an app-wide password. The app-wide
password is used to decode an encryption key, which itself is used
to encode and decode DEX account private keys. Wallets still have
their own individual passwords. This PR does not implement encryption,
but does suggest an interface for the upcoming work.

UI: The /register page's single form is broken into a series of forms,
the display of which is based on the current user state. The page walks
the user through setting an app-wide password, connecting to a Decred
wallet, and then creating a DEX account.