client/comms: refactoring
wsConn reconnect tweaks and logging

supervise all goroutines, cleanup

Remove unused wsConn.sendCh.

Allow (*wsConn).MessageSource to return the channel repeatedly, and
document this.  Also close this channel when wsConn shutsdown to signal
to the consumer that the conn is dead (if they didn't kill it).

wsconn.go: only queue reconn in Connect and keepAlive