db: Add MatchData.Active field and active column
The matches table has a status column with values corresponding to
order.MatchStatus, but this did not indicate if swap negotation was
still active.  A failed and revoked match could not have MatchComplete
as status, so a separate active flag is introduced to handle this case.
The active column also indicates matches with status MatchComplete,
which have both redeem txns fully confirmed, but which have not yet
received redeem acks from both counterparties.

The MatchArchiver.ActiveMatches now returns matches for a user
according to the active flag rather than status <> MatchComplete.
This helps the server and user identify which matches have pending
user actions.