client/comms: refactoring and bug fixes
wsConn reconnect tweaks and logging.

supervise all goroutines, using an internal context and cancellation function.

Remove unused wsConn.sendCh field.

Allow (*wsConn).MessageSource to return the channel repeatedly, and
document this. Also close this channel when wsConn shuts down to signal
to the consumer that the conn is dead (if they didn't kill it).

Only queue reconnection in keepAlive, not connect or Connect. Previously
the number of reconnects queued would grow exponentially because
connect() was queuing reconnect on Dial failure.