auth: add latestOrders, a list of a user's latest orders
latestOrders manages a list of the latest orders for a user. Its purpose
is to track cancelation frequency.  Orders are added with add, and total
and cancel counts are obtained with count.

Add ordsByTimeThenID, which implements sort.Interface, for sorting
an ord slice in ascending order by time and then order ID. This puts the
oldest order at the front and latest at the back of the slice.

(*latestOrders).add using sort.Search

auth: add RecordCompletedOrder and RecordCancel, tests

In cancel.go, store the targeted order id for cancel orders.

swap: use RecordCompletedOrder when redeem acks are received