db: SaveRedeemA also stores the secret
The secret is also stored in the matches table.

Although the secret is used by both redeem txns, and both are validated,
SaveRedeemA stores the secret.  Thus, SwapData now includes a
RedeemASecret field for the secret since A revealed the secret even if
the server heard about the secret from B's 'redeem' request first.

pg: refactor swap data storage functions to use the new
(*Archiver).updateMatchStmt method which is generalized for any SQL
statement with any number of arguments, expecting to update a single
row of the matches table for a market.