multi: connect order book
app:

Completes the `MarketsPage` js class, and hooks up the /markets page.
Better handling of zoom in the depth chart. Epoch orders are also
added to depth, but I plan to follow up with a customization to
differentiate the display of order book and epoch depth.

core:

Subscribe to the order book when the user navigates to the /market
page. Unsubscribe from the order 1 minute after last tab navigates
away from the market's page. Send notifications for book, unbook and
epoch orders. Send a notification when a new epoch is known to have
started, which could be triggered by either a order book notification
or a match_proof notification.

client/orders:

Modifies client/orders to be more permissive of sequencing issues,
which might be temporary, and also to allow orders from multiple epochs
to be in the queue at the same time. The latter is necessary because
order book notifications might come before the match_proof notification
for the prior epoch.