dexcctl: Allow secure input of arguments.
Add an option to bring up a secure terminal for input of sensitive
arguments that are not echoed. Arguments are appended to command line
arguments.

Prompt for passwords automatically.

Allows dexcctl to prompt for password arguments without any extra flags
from the user. Help messages have been divided up further to enable them
to be kept in the same location but have two ways of calling. They can
be called with password arguments included or not. Help without password
arguments is the default.

Separate args and password args.

So as to rely less on position, separate args when feeding to the
parser.

Only accept RawParams.

To lay the burden of correctness on the server, move all argument
parsing to the server. dexcctl, or any other client, will only construct
and send a RawParams as their msgjson payload and the server will parse
them and return a help message if there is a problem.

Handlers use parsers directly.

Because parsing has been moved to the server, parsing can be handled
individually by handlers and concrete types can be used.