core/webserver: client notification system
* core/webserver: client notification system

Adds a client notification system. core defines the interface, but
the methods are implemented by the db.Notification type. core defines
concrete types that may or may not carry auxiliary information beyond
the persistent db data that can be used for updating the browser
app state.

Notification above a certain "severity" are stored in the database.
(*Core).Login now returns a list of the 10 most recent notifications
for the browser to cache.

The browser uses Window.localStorage to store the most recent
notifications and prevent extraneous requests when multiple tabs
are used. As a demonstration of the usage of notifications in the
browser, the current order table on the markets page is now updated
live.