core: reload active orders and matches
Reload and prepare active orders and orders with active matches,
and the matches, from the database. Resolve the matches reported
in the server's 'connect' response with the local state. Only issues
errors when discrepancies are found for now, but more sophisticated
handling will be needed.

Unlocks wallets needed for active orders when the user signs in.