server/admin: Add webserver
Add an admin interface that uses tls to listen on an address for server
administration commands. WebAdminPass must be set. Uses the dex
server's cert and key files.

Do not read the password from a config file and only store the hash of
the password. Fail fast on an empty password.