dcrdex: use encrypt.Crypter for the signing key storage
Replace the unencrypted "dexprivkey" file with an encrypted "sigkey"
file. The sigkey file is a binary encoding of the signing key and the
encryption parameters, created using encode.BuildyBytes and
encode.Crypter:

    encode.BuildyBytes{0}.AddData(keyParams).AddData(encKey)

server/admin: Add PasswordHashPrompt, and modify PasswordPrompt
to retrieve the raw password bytes, not the hash of it.