webserver: data race bandaid
This supervises marketSyncer and goroutines. A *sync.WaitGroup field is
added to WebServer to facilitate this.  It is passed to newMarketSyncer.

This also copies the logger when starting goroutines, although this
should now be unnecessary if only one WebServer instance exists.

Ultimately the logger should either be made a field of WebServer and passed
around to other functions it calls, or remain a package-level variable
but do not set it in New (WebServer) constructor.