pg: cancel order matches are inactive
When storing a match where the taker is a cancel order, store with
status MatchComplete, and without the maker swap address. The match
will also be flagged as inactive since there is no associated swap
negotiation.  ActiveMatches can now never return such a match.

InsertMatch identified cancel orders for this handling by the taker's
order.  If the order.Match.Taker has a nil Trade, the match is inserted
with the handling described above.

Internally, this allows both maker and taker addresses to remain NULL in
the matches table.  Scan code is modified to use sql.NullString.