client: authDEX on reconnect.
Re-auth connections found in the conns map.

Make reconnect intervals increase incrementally so that a user may not
have to wait a full minute for a reconnect.

Give comms a logger.

Add notifications.

In order to give more information to the user, send notifications on
signAndRequest failures in Trade and Cancel. Also send a new SystemNote
on reconnections.

Cap the reconnect interval at one minute.

Add connect and disconnect notifications.

In order to inform the user whether the DEX the are trying to negotiate
on is currently connected, add notifications and a connected icon.

Show icon when disconnected.

Use core's Exchanges rather than user's.

A user's Exchanges are not updated when a connect/disconnect happens. In
order to serve fresh data to the webserver, send the core's Exchanges
data instead.

Refresh user on connect event.