dex/order: eliminate uid string cache in Prefix
Both id and uid caching are not thread safe, and the UID is easily
forgotten.  This removes the uid cache field, and modifies the UID
methods to always call OrderID's String method.

The Order ID field still gets computed and cached on first call to ID().

market: stamp the order a little earlier, fixing a rare Warnf panic

Regardless of order.Order implementation, the server time stamp is
required for order ID computation by definition of the order ID.
However, it is unfortunate that the implementations' ID caching is not
thread safe, requiring the first consumer to request the ID so it gets
written to the Order's cache before passing the order to other
goroutines that may cause a concurrent write.

The first action on receiving the order in Market.Run is to time stamp
it with the server time and process it with Market.processOrder where
it's ID is computed.