client/asset: Add ltc.
Tweak WalletConfig to include a couple of fields that are different for
litecoin. MinNetworkVersion is lowered to include litecoin's version as
offered on their website. Building latest from source may not be
possible for everyone. Include walletInfo in the config as it is
different from btc.

In order for bitcoin's chaincfg params to use new params correctly, they
must be registered.