swap: swap state
Swapper.matches map keys on MatchID instead of string

Rewrite (*Swapper).Run with graceful shutdown and state saving.

Use UnspentContract in processInit; Contract in setState.
Use Redemption in processRedeem; Input in setState.

auth: pendingRequest and RequestWhenConnected

Penalize flags connected users as suspended, but still attempts to close
accounts in the DB even if they are not found in the connected clients map.

handleConnect permits closed accounts to connect for dealing with active
swaps, but sets the clientInfo.suspended flag.

Refine application of RequestWhenConnected to decrement the connect
timeout on each successive reconnect so the user cannot keep
reconnecting and disconnecting forever.

Set DefaultConnectTimeout to 10 minutes.

swap,market: Add RequestWithConnected to stubs.

swap: implement coin waiter restore from State

Use RequestWhenConnected instead of RequestWithTimeout in both
processInit and processRedeem, which are called from the waiter
when the coins are found.

Implement a connection grace period on resume that increases penalty
timeout to connect timeout instead of the usual response timeout.

Deal with penalization in a separate AfterFunc (outside of the request's
expire function) so that penalization may still be applied without
cancelling the request.  This is important to enfoce community conduct
while still allowing the users to complete swaps since that is in
everyone's interest.  For this reason, also use RequestWhenConnected
in Negotiate even though there is no grace period there on restore.

TODO: figure out sensibility of penalization to revoke ack inaction.

swap: fix processAck not recognizing revoke acks

TODO: do something with the revoke ack sigs!!!

msgjson: rename and export signable => Signature for gob

auth,swap,market: pendingMessages and SendWhenConnected

(*AuthManager).SendWhenConnected is useful to help any message to
succeed in the event of transient connectivity issues, such as: new order
responses (market), successful verification of contracts or redeems
(swap), and any error responses.  But it is essential when resending
these messages on Swapper load with a saved state when users are
unlikely to be connected immediately after startup.

swap: remove (*Swapper).coinNotFound, calling respondError directly
from processInit and processRedeem.  This is cleaner and it allows the
decoded coin ID to be included with the coin waiter's timout message.

a bunch of cleanup and docs in the swap package.

market: Use SendWhenConnected in (*Market).processOrder and respondError

For now, just use the expire function to log that the message did not
reach its intended recipient. Consider other actions in processOrder.

Define a separate market.DefaultConnectTimeout.