multi: Remove scheme from connection map
This allows core methods to be called with a variety of different
patterns. Because net/url.Parse and net.SplitHostPort are not built for
our specific needs, a new function addrHost is added. This function will
parse dex addresses in a predictable manner that always produces a
host:port pair or just a host. Using the method on an already parsed
host will produce the same host.