auth,comms: on send/request failure remove response handler
In (*AuthManager).RequestWithTimeout if client.conn.Request returns with
a non-nil error:
- delete the response handler from the client's handler map
- stop the expire func Timer so it does not fire
The caller of Request/RequestWithTimeout should check the error and act
accordingly, which may not be calling the expire func.

In (*wsLink).Request, do the same handler and timer clean up if Send
returns a non-nil error. The caller (often authMgr.RequestWithTimeout)
will handle the error as appropriate.