asset: lift unspent requirements
The Contract interface now has a SwapAddress methoda.

The Backend.Contract method does not enforcing the output to be unspent.

In both dcr and btc backends, keep the UTXO type for FundingCoin, but we
will likely change FundingCoin to work with an Output type and instead
check spendability by the caller according to their requirements.

asset/btc: set maturity in newTXIO

asset/dcr: note that maturity cannot be set in newTXIO and requires a
fix to check the tag of the output index in question (e.g. ticket change
matures right away, ticket outputs are not spendable except by votes
and revokes, etc.)  This is a TODO for a follow-up fix.

asset: remove Backend.Input method and impls