app: add market list filter
This adds a small input at the top of the markets list that allows
filtering the displayed markets.

The market list used to be populated in the html template, and the
MarketsPage class would read its initial market list from there.
To be in line with patterns used elsewhere, the market list is now
populated by javascript using js template elements pulled and cloned
from the DOM.