core: fix locktime handling
This fixes locktime was not being set correctly.

The server intentionally sets the msgjson.Match.ServerTime based on the
match epoch so that locktime can be based off of a common point for both
maker and taker. Plus, taker wasn't setting their locktime correctly when
constructing the contract, which is now fixed.