core: better handle init request errors
Currently, init request errors prevent swap details from being saved
for a even though the swap tx has been broadcasted. This made it
impossible to re-attempt sending the init request after some time or
to refund the swap if it becomes necessary to do so.

This is corrected by saving relevant swap details to database even
if sending the init request fails. Also, if sending the init req fails for a
match, the request is re-tried in the next call to trade.tick().

Finally, if sending the init request does not complete successfully
before the swap's locktime expires, the swap is refunded.