dex,market,swap: unbookHook for swapper
It was a mistake for Swapper.checkInaction to revoke orders in
persistent storage, as they would have still been on Market's
in-memory book.

When a match fails, there is no good way to unbook the faulty order.
This creates a Swapper unbookHook function set on construction time
via Config.  In the dex manager, this unbook hood is a simple dispatcher
to the appropriate Market's new Unbook method.

Add (*Market).Unbook to remove the order from the in-memory book,
and if it was removed, unlock the coins, change the order's status to
revoked in storage via RevokeOrder, record the cancel with the
AuthManager via RecordCancel for cancellation ratio tracking, and
finally send a signal to the book router of the unbookAction.
The Unbook method does not require the Market to be running.

This also changes the shutdown behavior of Market.Run so that it does
not close the order feeds on shutdown.  This permits unbook
notifications to reach the book router after the Market has shutdown.
This also causes the book router to stay running after Market suspend,
as a closed channel stops the book router. The book router now only
stops on context cancellatoin.

Add Market.FeedDone so the book router can release the feed channel,
but this should not be necessary for the book router to do since the
book router will be shutdown after the market when the dex manager
properly orders the shutdown sequence.

server/swap: use RequestWhenConnected with revoke_match