server/asset/dcr: rename CoinDetails => FeeCoin, and restrict script
CoinDetails was only used to check fee payment, and it is required to
restrict the pkScript to non-stake P2PKH (single address).  These
restrictions make it not a genral purpose "coin details" function, so it
should not be named this way.

TODO: consider changing AuthManager.checkFee to accept the required
address, value, and confirmations rather than checking in the caller,
(*AuthManager).validateFee, since the checkFee/FeeCoin function already
restricts the coin's pkScript so much that it is not general purpose.