client/db/bolt: fix data corruption from reused buffers
The returned byte slice from bbolt.(*Bucket).Get is not valid outside of
the transaction where it is used.  This change fixes several places
where the []byte was used without copying it, resulting in corrupted
data.

Add the client/db/bolt.getCopy function to load a value and make a copy.
If the loaded value is an empty slice, it returns nil.  Thus, to
identify a non-existent value, use Bucket.Get directly instead.
getCopy covers most of the cases where we want to get the value in a
slice that is safe to use outside of a transaction, and when we do not
wish to load an empty non-nil slice when nil was actually stored.

Notably, this fixes decodeOrderBucket, and possibly filteredMatches,
makeWallet, and NotificationsN.