client/core,dex/order: fix incorrect fill amount when missing inactive matches
When dexc is restarted with a partially filled order where the previous
match(es) that partially filled the order are inactive, the fill amount
computation in (*trackedTrade).negotiate was incorrect.

This change adds the filled quantity for new matches rather than trying
to recompute and set the whole match amount.

(*trackedTrade).negotiate now attempts to process all matches even if
one of them fails for some reason.