client/comms: marshal before attempting link write
In client/comms.(*wsConn).Send, marshal the Message first so that we
do not send junk to the peer even if it fails to marshal completely,
which WriteJSON does.

Test marshal error with Send/Request, and removal of response handler
in this case with Requst.

Test RequestWithTimeout expiration handler and response handler removal.