multi: Add penalties table.
Because banning a user will become more complicated than just banned/not
banned, the banned information we hold for a user must be more detailed.
The best way to organize this new data is with a new table that holds
penalties. A combination of these penalties will decide wether or not a
user is banned.

Fields that are stored are:
 1. AccountID - the user involved
 2. BrokenRule - the broken rule
 3. Time - the time the penalization was enforced
 4. Duration - the length of penalization
 5. Details - a description of the infraction
 6. Forgiven - whether this penalization has been forgiven by the
    operator

Every Penalty is keyed by an incrementing id that is decided by the
database upon insertion.