client/core: use active order check in CloseWallet
(*dexConnection).hasOrders, which was used in (*Core).CloseWallet,
did not consider status of the order or its matches. CloseWallet should
permit locking the wallet if none of the orders/matches pertaining to
that wallet are active as defined by (*trackedTrade).isActive.

Although the dexConnection's trackedTrades are now periodically cleaned
out, this does the active check in CloseWallet too.

Nothing else required (*dexConnection).hasOrders, so it is simply
replaced with the new hasActiveAssetOrders method that is just like
hasActiveOrders, but filters by asset ID.