multi: Move websocket.go to separate package.
Because both the RPC and WEB servers use virtually the same websocket
server, move it to its own package. Tests are also split and moved.

The new server has a separate required client core interface that is
satisfied by core. The server is now created and started before creating
the WEB server and passed to it. It is now tied to the life of the core,
rather than the WEB server.