client: clean up inactive trades
* client/{asset,core}: return contract expiry from LocktimeExpired

* client/core: filter and retire inactive trades

Update (*trackedTrade).isActive to exclude refunded matches, and some
revoked matches depending on match status and side.
For revoked matches:
 - NewlyMatched requires no further action from either side
 - MakerSwapCast requires no further action from the taker
 - (TakerSwapCast requires action on both sides)
 - MakerRedeemed requires no further action from the maker
TODO: Consider a new dex/order.MatchRefunded MatchStatus. For now we
check the MatchProof.RefundCoin.  Add a comment in dex/order's
MatchStatus enum, and in (*trackedTrade) refundMatches where this status
would be set.

In the main trade ticker, clean out inative trades from the
dexConnection's trackedTrades map.

* client/core: use active order check in CloseWallet

(*dexConnection).hasOrders, which was used in (*Core).CloseWallet,
did not consider status of the order or its matches. CloseWallet should
permit locking the wallet if none of the orders/matches pertaining to
that wallet are active as defined by (*trackedTrade).isActive.

Although the dexConnection's trackedTrades are now periodically cleaned
out, this does the active check in CloseWallet too.

Nothing else required (*dexConnection).hasOrders, so it is simply
replaced with the new hasActiveAssetOrders method that is just like
hasActiveOrders, but filters by asset ID.

* only log contract expiry once in refundable check

* client/db/bolt: mirror active match checks in DEXOrdersWithActiveMatches

* client/core: record tick interval in dexConnection

Set the tickInterval on dexConnection construction, and use it when:
(1) creating the ticker in listen, and (2) when deciding to log the
contract expiry in (*trackedTrade).isRefundable