client/asset/{btc,dcr}: enable split funding transactions
Funding trades involves locking up coins (utxos w/ dcr and btc), which may be much
larger than the trade size. To avoid locking up excessive funds, which could go to
other orders, this allows the user to prepare outputs sized appropriately for their
orders.

This adds a setting to individual wallet configurations, txsplit, that enables this.
Description of the setting:

"When placing an order, create a \"split\" transaction to fund the order without
locking more of the wallet balance than necessary. Otherwise, excess funds
may be reserved to fund the order until the first swap contract is broadcast
during match settlement, or the order is canceled. This an extra transaction
for which network mining fees are paid. Used only for standing-type orders,
e.g. limit orders without immediate time-in-force."

This will help free up funds for more orders, thus creating a healthier order book
and a better user experience, but at the expense of an extra transaction.