client/asset/{btc,dcr}: fix locking issues
btc: resolve deadlock in FundOrder, no longer lock mtx in split()

dcr:
- Fix Swap unguarded lockFundingCoins call, also keep consistent view
- make unexported returnCoins that does not lock, for Swap.
- split() was not locking fundingMtx  before calling lockFundingCoins
- split() now keeps a consistend funding coin view from tx to locking  funds and split funds

Note that dcr.split() is thread-safe, while btc.split() is not. This is
to accommodate the different usage patterns in FundOrder.
The approach in btc is slightly more efficient because the map and
wallet lock are only updated once regardless of split or not.
While in dcr the initial funding updates the map and locks in the
wallet, and does it again if split() is used.